+++ model_fetch_flutter/lib/model_fetch_flutter.dart
@@ -1,2 +1,3 @@
+export 'src/bloc_auto_load_more.dart';
 export 'src/collection_bloc_builder.dart';
-export 'src/load_more.dart';
+export 'src/lazy_load_bloc_trailing.dart';